// src/types/contributions-received.d.ts
export interface ContributionsReceived {
  name_of_contributor: string;
  payment_type: string;
  city: string;
  state: string;
  zip: string;
  id_number: string;
  employer: string;
  occupation: string;
  amount: string;
  transaction_date: string;
  filed_date: string;
  transaction_number: string;
}